+++ docker/services/ruby/apps/video_describer/video_describer_app.rb
@@ -1,3 +1,3 @@
-class VideoDescriberOpenAI < MonadicApp
+class VideoDescriberApp < MonadicApp
   include VideoAnalyzeAgent
 end